Practicing nested ifs
This is just a simple calculator of 2 numbers where you input the first and second number and operator that will be used. Some of the input the user may enter would be the word of the operator, the operator sign or nothing at all. This code is a practice in producing an output using different scenarios. hehe